{"schema":"decodex/commit/1","summary":"Split protocol event state modules","authority":"manual"}
