+++ apps/decodex/src/state/protocol_events/hash.rs
@@ -1,0 +1,13 @@
+use sha2::{Digest as _, Sha256};
+
+pub(super) fn protocol_event_payload_sha256(payload: &str) -> String {
+	let digest = Sha256::digest(payload.as_bytes());
+	let mut hash = String::with_capacity(64);
+
+	for byte in digest {
+		hash.push(char::from(b"0123456789abcdef"[(byte >> 4) as usize]));
+		hash.push(char::from(b"0123456789abcdef"[(byte & 0x0f) as usize]));
+	}
+
+	hash
+}
